Finish refactoring cumulative metric logic
