+++ backend/server/graphql/calculations.py
@@ -1,6 +1,7 @@
 """Module for shared logic for calculating model metrics."""
 
-from typing import Optional, List
+from typing import Optional, List, Dict, Any
+from functools import partial
 
 from django.db.models import (
     Case,
@@ -19,6 +20,8 @@ from django.db.models import (
     QuerySet,
 )
 from django.utils import timezone
+import pandas as pd
+import numpy as np
 
 from server.models import TeamMatch
 
@@ -37,6 +40,11 @@ CUMULATIVE_METRICS_VALUES = [
     "predicted_winner__name",
     "predicted_win_probability",
 ]
+ROUND_NUMBER_LVL = 0
+GROUP_BY_LVL = 0
+# For regressors that might try to predict negative values or 0,
+# we need a slightly positive minimum to not get errors when calculating logarithms
+MIN_LOG_VAL = 1 * 10 ** -10
 
 
 def _calculate_cumulative_accuracy():
@@ -84,7 +92,7 @@ def _calculate_tip_points():
 
 def cumulative_metrics_query(
     prediction_query_set: QuerySet, additional_values: Optional[List[str]] = None,
-):
+) -> List[Dict[str, Any]]:
     """
     Chain methods onto a Prediction query set to calculate cumulative model metrics.
     """
@@ -107,4 +115,98 @@ def cumulative_metrics_query(
             cumulative_accuracy=_calculate_cumulative_accuracy(),
         )
         .values(*values_args)
+    )
+
+
+def _filter_by_round(data_frame: pd.DataFrame, round_number: Optional[int] = None):
+    if round_number is None:
+        return data_frame
+
+    round_number_filter = (
+        data_frame.index.get_level_values(ROUND_NUMBER_LVL).max()
+        if round_number == -1
+        else round_number
+    )
+
+    return data_frame.loc[(round_number_filter, slice(None)), :]
+
+
+def _calculate_cumulative_bits(data_frame: pd.DataFrame):
+    return (
+        data_frame.groupby("ml_model__name")
+        .expanding()["bits"]
+        .sum()
+        .rename("cumulative_bits")
+        .reset_index(level=GROUP_BY_LVL, drop=True)
+    )
+
+
+# Raw bits calculations per http://probabilistic-footy.monash.edu/~footy/about.shtml
+def _calculate_bits(data_frame: pd.DataFrame):
+    positive_pred = lambda y_pred: (
+        np.maximum(y_pred, np.repeat(MIN_LOG_VAL, len(y_pred)))
+    )
+    draw_bits = lambda y_pred: (
+        1 + (0.5 * np.log2(positive_pred(y_pred * (1 - y_pred))))
+    )
+    win_bits = lambda y_pred: 1 + np.log2(positive_pred(y_pred))
+    loss_bits = lambda y_pred: 1 + np.log2(positive_pred(1 - y_pred))
+
+    return np.where(
+        data_frame["match__margin"] == 0,
+        draw_bits(data_frame["predicted_win_probability"]),
+        np.where(
+            data_frame["match__winner__name"] == data_frame["predicted_winner__name"],
+            win_bits(data_frame["predicted_win_probability"]),
+            loss_bits(data_frame["predicted_win_probability"]),
+        ),
+    )
+
+
+# TODO: I've migrated the simple calculations over to SQL, but calculations
+# based on margin_diff are difficult, because Django doesn't allow an annotation
+# based on an aggregation (margin_diff uses Max/Min), and `Window` can't be used
+# in an `aggregate` call. I may need to resort to raw SQL, but that would probably
+# still require figuring out why the ORM doesn't like this combination.
+def _calculate_cumulative_mae(data_frame: pd.DataFrame):
+    return (
+        data_frame.groupby("ml_model__name")
+        .expanding()["absolute_margin_diff"]
+        .mean()
+        .round(2)
+        .rename("cumulative_mean_absolute_error")
+        .reset_index(level=GROUP_BY_LVL, drop=True)
+    )
+
+
+def _calculate_cumulative_margin_difference(data_frame: pd.DataFrame):
+    return (
+        data_frame.groupby("ml_model__name")
+        .expanding()["absolute_margin_diff"]
+        .sum()
+        .rename("cumulative_margin_difference")
+        .reset_index(level=GROUP_BY_LVL, drop=True)
+    )
+
+
+def calculate_cumulative_metrics(
+    metric_values: List[Dict[str, Any]], round_number: Optional[int]
+) -> pd.DataFrame:
+    """Calculate cumulative methods that can't be calculated via the ORM."""
+    return (
+        pd.DataFrame(metric_values)
+        # We fill missing win probabilities with 0.5, because that's the equivalent
+        # of not picking a winner. 0 would represent an extreme prediction
+        # and result in large negative bits calculations.
+        .fillna({"predicted_win_probability": 0.5})
+        .fillna(0)
+        .assign(bits=_calculate_bits)
+        .assign(
+            cumulative_margin_difference=_calculate_cumulative_margin_difference,
+            cumulative_bits=_calculate_cumulative_bits,
+            cumulative_mean_absolute_error=_calculate_cumulative_mae,
+        )
+        .groupby(["match__round_number", "ml_model__name"])
+        .last()
+        .pipe(partial(_filter_by_round, round_number=round_number))
     )
