+++ backend/server/graphql/schema.py
@@ -1,7 +1,6 @@
 """GraphQL schema for all queries."""
 
-from typing import List, Optional
-from functools import partial
+from typing import List
 
 import graphene
 from django.utils import timezone
@@ -12,7 +11,10 @@ import numpy as np
 
 from server.models import Prediction, Match, MLModel
 from server.types import RoundMetrics
-from server.graphql.calculations import cumulative_metrics_query
+from server.graphql.calculations import (
+    cumulative_metrics_query,
+    calculate_cumulative_metrics,
+)
 from .types import (
     SeasonType,
     PredictionType,
@@ -35,13 +37,6 @@ RoundModelMetrics = TypedDict(
     "RoundModelMetrics",
     {"match__round_number": int, "model_metrics": pd.DataFrame, "matches": QuerySet},
 )
-
-
-ROUND_NUMBER_LVL = 0
-GROUP_BY_LVL = 0
-# For regressors that might try to predict negative values or 0,
-# we need a slightly positive minimum to not get errors when calculating logarithms
-MIN_LOG_VAL = 1 * 10 ** -10
 
 
 class RoundMetricsType(graphene.ObjectType):
@@ -82,80 +77,63 @@ class RoundMetricsType(graphene.ObjectType):
     )
 
     @staticmethod
-    def resolve_season(root: RoundMetrics, _info):
+    def resolve_season(round_metrics: RoundMetrics, _info):
         """Get season value from match__start_date_time."""
-        return root["match__start_date_time"].year
+        return round_metrics["match__start_date_time"].year
 
 
-def _filter_by_round(data_frame: pd.DataFrame, round_number: Optional[int] = None):
-    if round_number is None:
-        return data_frame
+def _consolidate_competition_model_metrics(model_metrics: pd.DataFrame) -> RoundMetrics:
+    assert model_metrics["ml_model__used_in_competitions"].all()
 
-    round_number_filter = (
-        data_frame.index.get_level_values(ROUND_NUMBER_LVL).max()
-        if round_number == -1
-        else round_number
+    principle_data = (
+        model_metrics.query("ml_model__is_principle == True")
+        .reset_index(drop=False)
+        .set_index("match__id")
+        # We replace previously-filled values with NaNs to make it easier to fill
+        # missing principle metrics with metrics from the other competition models.
+        # It's okay if we NaNify a legitimate 0/0.5, because the other model(s)
+        # will just fill the NaN with the same neutral value.
+        .replace(
+            to_replace={
+                "cumulative_mean_absolute_error": 0,
+                "cumulative_margin_difference": 0,
+                "cumulative_bits": 0,
+                "predicted_margin": 0,
+                "predicted_win_probability": 0.5,
+            },
+            value={
+                "cumulative_mean_absolute_error": np.nan,
+                "cumulative_margin_difference": np.nan,
+                "cumulative_bits": np.nan,
+                "predicted_margin": np.nan,
+                "predicted_win_probability": np.nan,
+            },
+        )
     )
 
-    return data_frame.loc[(round_number_filter, slice(None)), :]
-
-
-def _calculate_cumulative_bits(data_frame: pd.DataFrame):
-    return (
-        data_frame.groupby("ml_model__name")
-        .expanding()["bits"]
+    non_principle_data = (
+        model_metrics.query("ml_model__is_principle == False")
+        .fillna(0)
+        .replace(
+            to_replace={"predicted_win_probability": 0.5},
+            value={"predicted_win_probability": 0},
+        )
+        .select_dtypes("number")
+        .groupby("match__id")
+        # We can sum because each prediction type will only have one model with values,
+        # and the rest of the rows will be zeros.
         .sum()
-        .rename("cumulative_bits")
-        .reset_index(level=GROUP_BY_LVL, drop=True)
     )
 
+    consolidated_metrics = principle_data.fillna(non_principle_data).to_dict("records")
 
-# Raw bits calculations per http://probabilistic-footy.monash.edu/~footy/about.shtml
-def _calculate_bits(data_frame: pd.DataFrame):
-    positive_pred = lambda y_pred: (
-        np.maximum(y_pred, np.repeat(MIN_LOG_VAL, len(y_pred)))
-    )
-    draw_bits = lambda y_pred: (
-        1 + (0.5 * np.log2(positive_pred(y_pred * (1 - y_pred))))
-    )
-    win_bits = lambda y_pred: 1 + np.log2(positive_pred(y_pred))
-    loss_bits = lambda y_pred: 1 + np.log2(positive_pred(1 - y_pred))
-
-    return np.where(
-        data_frame["match__margin"] == 0,
-        draw_bits(data_frame["predicted_win_probability"]),
-        np.where(
-            data_frame["match__winner__name"] == data_frame["predicted_winner__name"],
-            win_bits(data_frame["predicted_win_probability"]),
-            loss_bits(data_frame["predicted_win_probability"]),
-        ),
+    assert len(consolidated_metrics) == 1, (
+        "Latest round predictions should be in the form of a single data set "
+        "composed of all competition models, but multiple sets were calculated:\n"
+        f"{consolidated_metrics}"
     )
 
-
-# TODO: I've migrated the simple calculations over to SQL, but calculations
-# based on margin_diff are difficult, because Django doesn't allow an annotation
-# based on an aggregation (margin_diff uses Max/Min), and `Window` can't be used
-# in an `aggregate` call. I may need to resort to raw SQL, but that would probably
-# still require figuring out why the ORM doesn't like this combination.
-def _calculate_cumulative_mae(data_frame: pd.DataFrame):
-    return (
-        data_frame.groupby("ml_model__name")
-        .expanding()["absolute_margin_diff"]
-        .mean()
-        .round(2)
-        .rename("cumulative_mean_absolute_error")
-        .reset_index(level=GROUP_BY_LVL, drop=True)
-    )
-
-
-def _calculate_cumulative_margin_difference(data_frame: pd.DataFrame):
-    return (
-        data_frame.groupby("ml_model__name")
-        .expanding()["absolute_margin_diff"]
-        .sum()
-        .rename("cumulative_margin_difference")
-        .reset_index(level=GROUP_BY_LVL, drop=True)
-    )
+    return consolidated_metrics[0]
 
 
 class Query(graphene.ObjectType):
@@ -297,77 +275,9 @@ class Query(graphene.ObjectType):
             prediction_query_set, additional_metric_values,
         )
 
-        metrics_df = (
-            pd.DataFrame(metric_values)
-            # We fill missing win probabilities with 0.5, because that's the equivalent
-            # of not picking a winner. 0 would represent an extreme prediction
-            # and result in large negative bits calculations.
-            .fillna({"predicted_win_probability": 0.5})
-            .fillna(0)
-            .assign(bits=_calculate_bits)
-            .assign(
-                cumulative_margin_difference=_calculate_cumulative_margin_difference,
-                cumulative_bits=_calculate_cumulative_bits,
-                cumulative_mean_absolute_error=_calculate_cumulative_mae,
-            )
-            .groupby(["match__round_number", "ml_model__name"])
-            .last()
-            .pipe(partial(_filter_by_round, round_number=max_match.round_number))
-        )
+        metrics_df = calculate_cumulative_metrics(metric_values, max_match.round_number)
 
-        assert metrics_df["ml_model__used_in_competitions"].all()
-
-        principle_data = (
-            metrics_df.query("ml_model__is_principle == True")
-            .reset_index(drop=False)
-            .set_index("match__id")
-            # We replace previously-filled values with NaNs to make it easier to fill
-            # missing principle metrics with metrics from the other competition models.
-            # It's okay if we NaNify a legitimate 0/0.5, because the other model(s)
-            # will just fill the NaN with the same neutral value.
-            .replace(
-                to_replace={
-                    "cumulative_mean_absolute_error": 0,
-                    "cumulative_margin_difference": 0,
-                    "cumulative_bits": 0,
-                    "predicted_margin": 0,
-                    "predicted_win_probability": 0.5,
-                },
-                value={
-                    "cumulative_mean_absolute_error": np.nan,
-                    "cumulative_margin_difference": np.nan,
-                    "cumulative_bits": np.nan,
-                    "predicted_margin": np.nan,
-                    "predicted_win_probability": np.nan,
-                },
-            )
-        )
-
-        non_principle_data = (
-            metrics_df.query("ml_model__is_principle == False")
-            .fillna(0)
-            .replace(
-                to_replace={"predicted_win_probability": 0.5},
-                value={"predicted_win_probability": 0},
-            )
-            .select_dtypes("number")
-            .groupby("match__id")
-            # We can sum because each prediction type will only have one model with values,
-            # and the rest of the rows will be zeros.
-            .sum()
-        )
-
-        consolidated_metrics = principle_data.fillna(non_principle_data).to_dict(
-            "records"
-        )
-
-        assert len(consolidated_metrics) == 1, (
-            "Latest round predictions should be in the form of a single data set "
-            "composed of all competition models, but multiple sets were calculated:\n"
-            f"{consolidated_metrics}"
-        )
-
-        return consolidated_metrics[0]
+        return _consolidate_competition_model_metrics(metrics_df)
 
     @staticmethod
     def resolve_fetch_ml_models(
